Avoid a kill(SIGTERM, -1)
git-svn-id: https://svn.smr.arpa.emr.it/arkimet/arkimet/trunk@2244 9b5092fc-ae44-42f6-b796-0d08dad2e17b

+++ arki/postprocess.cc
@@ -197,8 +197,11 @@ struct Child : public utils::IODispatcher
 
     void terminate()
     {
-        cmd.kill(SIGTERM);
-        cmd.wait();
+        if (cmd.pid() != -1)
+        {
+            cmd.kill(SIGTERM);
+            cmd.wait();
+        }
     }
 };
